pyproject.toml
[tool.poetry]
package-mode = false
name = "inf-3dp"
version = "0.1.0"
description = "inf-3dp"
authors = ["QU Jiasheng <jsqu@mae.cuhk.edu.hk>"]
license = "mit"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.15"
plyfile = "^1.1.3"
potpourri3d = "^1.3"
pyqt5 = "^5.15.11"
pyvistaqt = "^0.11.3"
scikit-image = "^0.25.2"
torch-tb-profiler = "^0.4.3"
trimesh = "^4.9.0"
configargparse = "^1.7.1"
tqdm = "^4.67.1"
h5py = "^3.15.1"
pyyaml = "^6.0.3"
torch = "^2.3.0"
torchvision = "^0.18.0"
torchaudio = "^2.3.0"


[[tool.poetry.source]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
